+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K62KZ6CKP62NJQE7X8Q9SFKH"
+        "01K62YGYCM426DXKA2WRZX1V7R"
     )
     r"""The id of the resource"""