+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -578,7 +578,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K62KZ68TAT5EHKJKEJ1QWV1V"
+        "01K62YGY99XT5HTD1YVGV5G48D"
     )
 
     requires_approval: Optional[bool] = False